Add authentication using JWT

+++ src/Service/PersonService.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 namespace App\Service;
 
 use App\Entity\Person;
+use Doctrine\DBAL\Exception\UniqueConstraintViolationException;
 use Doctrine\ORM\EntityManagerInterface;
 use Ramsey\Uuid\UuidInterface;
 
@@ -37,9 +38,18 @@ final class PersonService implements PersonServiceInterface
         $person->setBirthDate(\DateTime::createFromFormat('d/m/Y', $payload['birth_date']));
 
         $this->entityManager->beginTransaction();
-        $this->entityManager->persist($person);
-        $this->entityManager->flush();
-        $this->entityManager->commit();
+
+        try {
+            $this->entityManager->persist($person);
+            $this->entityManager->flush();
+            $this->entityManager->commit();
+        } catch (UniqueConstraintViolationException $exception) {
+            $this->entityManager->rollback();
+            throw new \Exception('Email is already registered.');
+        } catch (\Exception $exception) {
+            $this->entityManager->rollback();
+            throw new \Exception('Failed to register person.');
+        }
 
         return $person;
     }
@@ -48,6 +58,13 @@ final class PersonService implements PersonServiceInterface
     {
         $person = $this->entityManager->getRepository(Person::class)
             ->find($uuid);
+
+        if (is_null($person)) {
+            throw new \Exception(
+                sprintf('Person not found with id "%s"', $uuid->toString())
+            );
+        }
+
         $person->setName($payload['name']);
         $person->setEmail($payload['email']);
         $person->setBirthDate(\DateTime::createFromFormat('d/m/Y', $payload['birth_date']));
@@ -64,6 +81,13 @@ final class PersonService implements PersonServiceInterface
     {
         $person = $this->entityManager->getRepository(Person::class)
             ->find($uuid);
+
+        if (is_null($person)) {
+            throw new \Exception(
+                sprintf('Person not found with id "%s"', $uuid->toString())
+            );
+        }
+
         $this->entityManager->beginTransaction();
         $this->entityManager->remove($person);
         $this->entityManager->flush();
